stage01 1-4 set

+++ code/stage01/1-3-map/StudentLookup.h
@@ -2,12 +2,12 @@
 #include <string>
 #include <unordered_map>
 
-// 练习2 学生姓名->分数，按姓名查询
+// 练习5 学生姓名->分数，按姓名查询
 class StudentLookup {
 public:
-    void add(const std::string& name, int score);
-    int query(const std::string& name) const;
+    void add(const std::string& name, double score);
+    double query(const std::string& name) const;
 
 private:
-    std::unordered_map<std::string, int> records_;
+    std::unordered_map<std::string, double> records_;
 };
